refactor: disambiguate the module names the decomposition introduced
- Keep `rule.rs` a flat file beside the `rules/` directory rather than a parallel split directory.
- Rename `colon_targets/docstring.rs` to `docstring_args.rs`, distinguishing it from the `primitives/docstring/` module.
